+++ static/js/components/tiles/chart_tile.tsx
@@ -18,19 +18,18 @@
  * A container for any tile containing a chart.
  */
 
-import React, { useContext, useRef, useState } from "react";
+import React, { useRef } from "react";
 
 import { ASYNC_ELEMENT_HOLDER_CLASS } from "../../constants/css_constants";
 import { INITAL_LOADING_CLASS } from "../../constants/tile_constants";
 import { ChartEmbed } from "../../place/chart_embed";
-import { NlSessionContext } from "../../shared/context";
-import { onChartThumbDownClick } from "../../utils/nl_interface_utils";
 import {
   formatString,
   getChartTitle,
   getMergedSvg,
   ReplacementStrings,
 } from "../../utils/tile_utils";
+import { NLChartFeedback } from "../nl_feedback";
 import { ChartFooter } from "./chart_footer";
 interface ChartTileContainerProp {
   id: string;
@@ -50,10 +49,8 @@ interface ChartTileContainerProp {
 }
 
 export function ChartTileContainer(props: ChartTileContainerProp): JSX.Element {
-  const nlSessionId = useContext(NlSessionContext);
   const containerRef = useRef(null);
   const embedModalElement = useRef<ChartEmbed>(null);
-  const [isThumbClicked, setIsThumbClicked] = useState(false);
 
   // on initial loading, hide the title text
   const title = !props.isInitialLoading
@@ -83,25 +80,7 @@ export function ChartTileContainer(props: ChartTileContainerProp): JSX.Element {
         sources={props.sources}
         handleEmbed={showEmbed ? handleEmbed : null}
       />
-
-      {nlSessionId && (
-        <div className="nl-feedback">
-          <span
-            className={`thumb-down ${isThumbClicked ? "thumb-dim" : ""}`}
-            onClick={() => {
-              return onChartThumbDownClick(
-                props.id,
-                nlSessionId,
-                isThumbClicked,
-                setIsThumbClicked
-              );
-            }}
-          >
-            &#128078;
-          </span>
-        </div>
-      )}
-
+      <NLChartFeedback id={props.id} />
       {showEmbed && <ChartEmbed ref={embedModalElement} />}
     </div>
   );
