+++ static/js/components/tiles/top_event_tile.tsx
@@ -20,7 +20,7 @@
 
 import axios from "axios";
 import _ from "lodash";
-import React, { memo, useContext, useEffect, useRef, useState } from "react";
+import React, { memo, useEffect, useRef, useState } from "react";
 
 import {
   ASYNC_ELEMENT_CLASS,
@@ -29,7 +29,6 @@ import {
 import { INITAL_LOADING_CLASS } from "../../constants/tile_constants";
 import { formatNumber } from "../../i18n/i18n";
 import { ChartEmbed } from "../../place/chart_embed";
-import { NlSessionContext } from "../../shared/context";
 import { NamedPlace, NamedTypedPlace } from "../../shared/types";
 import {
   DisasterEventPoint,
@@ -42,9 +41,9 @@ import {
 } from "../../types/subject_page_proto_types";
 import { stringifyFn } from "../../utils/axios";
 import { rankingPointsToCsv } from "../../utils/chart_csv_utils";
-import { onChartThumbDownClick } from "../../utils/nl_interface_utils";
 import { getPlaceNames } from "../../utils/place_utils";
 import { formatPropertyValue } from "../../utils/property_value_utils";
+import { NLChartFeedback } from "../nl_feedback";
 import { ChartFooter } from "./chart_footer";
 
 const DEFAULT_RANKING_COUNT = 10;
@@ -66,8 +65,6 @@ interface TopEventTilePropType {
 export const TopEventTile = memo(function TopEventTile(
   props: TopEventTilePropType
 ): JSX.Element {
-  const nlSessionId = useContext(NlSessionContext);
-  const [isThumbClicked, setIsThumbClicked] = useState(false);
   const embedModalElement = useRef<ChartEmbed>(null);
   const chartContainer = useRef(null);
   const [eventPlaces, setEventPlaces] =
@@ -215,23 +212,7 @@ export const TopEventTile = memo(function TopEventTile(
           />
         </div>
       </div>
-      {nlSessionId && (
-        <div className="nl-feedback">
-          <span
-            className={`thumb-down ${isThumbClicked ? "thumb-dim" : ""}`}
-            onClick={() => {
-              return onChartThumbDownClick(
-                props.id,
-                nlSessionId,
-                isThumbClicked,
-                setIsThumbClicked
-              );
-            }}
-          >
-            &#128078;
-          </span>
-        </div>
-      )}
+      <NLChartFeedback id={props.id} />
       <ChartEmbed ref={embedModalElement} />
     </div>
   );
